# public/conceptionRessources/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "home-iot-api",
    "description": "The API for learning formation NodeJS project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/REMITAFFOREAU_1/LearningFormation/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
...
    "/vehicles/category/{categoryVehicleId}": {
      "get": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Get vehicles by category",
        "operationId": "findVehiclesByCategory",
        "parameters": [
          {
            "name": "categoryVehicleId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_1"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/detail/{id}": {
      "get": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Get vehicle details",
        "operationId": "getVehicleDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_2"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
...
    "/vehicles/update/{id}": {
      "put": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Update a vehicle",
        "operationId": "updateVehicle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400_2"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/delete/{id}": {
      "delete": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Delete a vehicle",
        "operationId": "deleteVehicle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401_1"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "r@yahoo.fr"
          },
          "pseudo": {
            "type": "string",
            "example": "coco"
          },
          "password": {
            "type": "string",
            "format": "hash",
            "example": "0729a580-2240-11e6-9eb5-0002a5d5c51b"
          },
          "role": {
            "type": "string",
            "example": "prestataire"
          }
        }
      },
      "CategoryVehicle": {
        "type": "object",
        "properties": {
          "libelle": {
            "type": "string",
            "example": "Automobile"
          }
        }
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "libelle": {
            "type": "string",
            "example": "Renault Mégane"
          },
          "description": {
            "type": "string",
            "example": "Berling rouge de grand standing"
          },
          "imageUrl": {
            "type": "string",
            "example": "http://1230.fr"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryVehicle"
          },
          "priceDay": {
            "type": "integer",
            "format": "int32",
            "example": 12
          },
          "prestataire": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "inline_response_400": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "inline_response_404_1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "No vehicle in this category."
          }
        }
      },
      "inline_response_404_2": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Vehicle not found.."
          }
        }
      },
...
      "inline_response_400_2": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Missing or invalid authentication token."
          }
        }
      },
      "inline_response_200_1": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Véhicle updated !"
          }
        }
      },
      "inline_response_200_2": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Vehicle deleted !"
          }
        }
      },
      "inline_response_401_1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Missing or invalid authentication token or prestataire is not the user."
          }
        }
      },
...
    }
  }
}